+++ pages/index.tsx
@@ -1,7 +1,7 @@
 import { GetStaticProps } from 'next'
 import { getPosts } from '../utils/mdx'
 import { PAGE_SIZE } from '../utils/constants'
-import Home, { HomeProps } from '../components/templates/HomeLayout'
+import Home, { HomeProps } from '../components/templates/Home'
 
 export const getStaticProps: GetStaticProps<HomeProps> = async (context) => {
   const posts = await getPosts({ limit: PAGE_SIZE })
